Convert an object of WP_Error class to an array
This commit is fixing a bug which leads to every successful remote request being turned into an array with `response => [ code => 500 ]` value and which is also allowing the `LFAPPS_Http_Extension::request` method to return an object of WP_Error.

This commit is fixing the bug by removing the negation (`!`) from conditional.

Work towards #111