Always return an array from LFAPPS_Http_Extension::request
The method is currently retunning a string or a boolean value while in WordPress.com VIP environment, while the rest of the codebase expects an array.

This commit is fixing the issue by replacing the `wpcom_vip_file_get_contents` function call by `vip_safe_wp_remote_get` and adding a logic for converting an object of WP_Error into an array.

Work towards #111